Make adapter inner class of ThrottledTaskRunner
